First draft of the project. Testing for compatibility of writeup to GitHub.
