+++ p2-NInt/ethanwhite/writeup/writeup.md
@@ -1,0 +1,29 @@
+---
+	Author: Ethan White
+	Topic: Numerical Integration Project
+	Course: TN Tech PHYS 4130
+	Term: Spring 2025 
+---
+
+### Introduction
+
+<figure>
+  <img src=Figure_1.png>
+  <figcaption>4x4 plot of first through fourth order polynomials and their products. </figcaption>
+</figure>
+<p>&nbsp;</p> 
+
+### Attribution
+
+The main components of this project were created in the homework assignments (up to simpson's rule, at least). Zeke found the first source, [1], helpful so I used it.
+
+1. https://matplotlib.org/stable/gallery/subplots_axes_and_figures/subplots_demo.html
+
+### Timekeeping
+
+This project took about 15-20 hours of work. The main struggles were within the extension questions.
+
+### Languages, Libraries, Lessons Learned
+	1. Python was used for the entirity of this project. 
+
+	2. I used numpy and math to.. do math stuff. I used matplotlib.pyplot to create the Legendre plot portion of the project. I used scipy to pull out the weights and roots of the Legendre polynomials.
